Add option to find python files in other dirs
By default we scan any package files for imports to check against
dependencies. However, sometimes one might be scanning interested in
files outside of the package (e.g. tests) which this option allows to
scan.

Note: this might be a bit redundant with editable installs, e.g. with a
structure like

    .
    ├── some_pkg
    │   └── __init__.py
    └── tests
        ├── __init__.py

Then if the `.pth` files points at the root we'll discover the tests
anyway....which is maybe a nuisance if one wanted to scan _only_
`some_pkg` for deps us.