Avoid `poetry` uninstalling dependencies
When a dependency was specified as an extra, but `poetry` wasn't
instructed to install any extras then it would remove the dependency. I
don't want this, so just force it to install all extras all the time.

If I start doing any fancier installs I might just move the install
scripts to within each distributions directory, but no needed for now