Generate `requirements.txt` via `pip-compile`
I want to ensure development env doesn't break unexpectedly when a new,
incompatible, version of a dependency is released. We should still be
aware of such breakages via `dependabot` trying to update things.

To test out dependabot this change explicitly uses an older version of
`packaging` so can see if dependabot does with that.

The file was generated via:

    $ pip-compile --generate-hashes --reuse-hashes --strip-extras --output-file requirements.txt