Update typing in test utils
Remove type ignore only needed for Python 3.7 or less, since we've
dropped support for Python 3.7.

Fix error on return type for stubbed out function, the error was:

    tests/utils.py:40: error: Return type "PathLike[str]" of "locate_file" incompatible with return type "Path" in supertype "Distribution"  [override]

This `SimplePath` isn't exported (but should be in the future[1]) but
it's just a protocol implementing some bits of `pathlib.Path`[2] so drop
that in instead.

[1] https://github.com/python/importlib_metadata/issues/494
[2] https://github.com/python/cpython/blob/69a4063ca516360b5eb96f5432ad9f9dfc32a72e/Lib/importlib/metadata/_meta.py#L47